fixed menu

+++ app/Http/Controllers/ContentController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\ResourceSaveRequest;
+use App\Models\Available;
 use App\Models\Content;
+use App\Models\Group;
 use App\Models\Tree;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -19,17 +21,17 @@ class ContentController extends Controller
      */
     public function saveImage(Request $request)
     {
-        if($request->file('image')){
-        $path   = Image::read($request->file('image'));
-        $resize = $path->scaleDown(1024, 1024)->toWebp(90);
-        $path = "contentImages/" . Auth::user()->id . "/" . Str::random(40) . ".webp";
-        Storage::disk("public")->put($path, $resize);
-        return response()->json(['success' => 1, 'file' => ['url' => URL::to('/') . "/" . $path]], 200);
-        }        
+        if ($request->file('image')) {
+            $path   = Image::read($request->file('image'));
+            $resize = $path->scaleDown(1024, 1024)->toWebp(90);
+            $path = "contentImages/" . Auth::user()->id . "/" . Str::random(40) . ".webp";
+            Storage::disk("public")->put($path, $resize);
+            return response()->json(['success' => 1, 'file' => ['url' => URL::to('/') . "/" . $path]], 200);
+        }
     }
     public function saveFile(Request $request)
     {
-        $path = Storage::disk("public")->putFile("contentFiles/".Auth::user()->id, $request->file('file'));
+        $path = Storage::disk("public")->putFile("contentFiles/" . Auth::user()->id, $request->file('file'));
         return response()->json(['success' => 1, 'file' => ['url' => URL::to('/') . "/" . $path]], 200);
     }
 
@@ -76,9 +78,11 @@ class ContentController extends Controller
             ]);
             $fileId = Content::create([
                 'tree_id' => $tree->id,
-                'accessibility' => $request->accessibility ?: false,
+                'accessibility' => $request->accessibility,
                 'data' => $request->data,
             ]);
+
+            $this->changeAvailables($tree->id, $request->availables, $request->accessibility);
 
             return response()->json(['success' => true, 'message' => 'Файл успешно создан', 'id' => $fileId->tree_id]);
         } else {
@@ -86,28 +90,69 @@ class ContentController extends Controller
             $fileId = Content::where("tree_id", $request->id)->first();
 
             $fileId->update([
-                'accessibility' => false,
+                'accessibility' => $request->accessibility,
                 'data' => $request->data,
             ]);
             $tree->update([
                 'name' => $request->name,
             ]);
+
+            $this->changeAvailables($tree->id, $request->availables, $request->accessibility);
+
             return response()->json(['success' => true, 'message' => 'Данные файла обновлены', 'id' => $fileId->tree_id]);
         }
     }
-    public function getResource($content)
+
+    public function changeAvailables($tree, $availables, $accessibility)
     {
 
+        Available::where('tree_id', $tree)->delete();
+
+        if (!$accessibility) {
+            foreach (json_decode($availables) as $available) {
+                if ($available->checked) {
+                    Available::create(
+                        [
+                            'group_id' => $available->id,
+                            'tree_id' => $tree,
+                        ]
+                    );
+                }
+            }
+        }
+    }
+
+    public function getResource($content)
+    {
         $tree = Tree::find($content);
         if (!$tree) {
             return response()->json(['success' => false, "message" => 'Файла не существует']);
         }
         $res = Content::with("tree")->where("tree_id", $content)->first();
-        return response()->json(['content' => $res, "name" => $tree->name]);
+
+        $availablesGroups = [];
+
+        $all = Group::get();
+        $g = Available::where('tree_id', $content)->pluck('group_id')->toArray();
+
+
+        foreach ($all as $value) {
+            $availablesGroups[] = ['id' => $value->id, 'name' => $value->name, 'checked' => in_array($value->id, $g)];
+        }
+
+        return response()->json(["name" => $tree->name, 'content' => $res,  'groups' => $availablesGroups]);
     }
 
-    public function delResource($content)
+    public function delResource($content, Request $request)
     {
+        if ($request->user()->role == 'user') {
+            return response()->json(["success" => false, 'message' => 'Недостаточно прав']);
+        }
+
+        if ($request->user()->role == 'ceo' and $request->user()->id != $content->user_id) {
+            return response()->json(["success" => false, 'message' => 'Недостаточно прав']);
+        }
+
         Tree::find($content)->delete();
         return response()->json(["success" => true, 'message' => 'Файл удален']);
     }
